refactor: split dashboard-api.ts into modular route files
- Extract layout and styles into separate files in /layout directory
- Create individual route files for better organization:
  - auth.ts - Login/logout routes
  - overview.ts - Main dashboard overview
  - requests.ts - Recent requests page
  - request-details.ts - Individual request details
  - token-usage.ts - Token usage dashboard (placeholder)
- Create pagination utilities in utils/pagination.ts
- Update imports to use new modular structure
- Fix conversation-detail.ts to import layout from new location
- Maintain backward compatibility with existing functionality

This improves code organization and maintainability by breaking down the
large 2500+ line file into smaller, focused modules.
