--- docker/docker-compose.local.yml
@@ -1,149 +1,0 @@
-# Docker Compose configuration for local development with locally built images
-# 
-# Usage from project root:
-#   ./docker-up.sh -f docker-compose.local.yml up -d
-# Or from docker directory:
-#   docker compose -f docker-compose.local.yml --env-file ../.env up -d
-#
-# This file uses locally built images instead of registry images
-
-services:
-  # PostgreSQL Database
-  postgres:
-    image: postgres:15-alpine
-    container_name: claude-nexus-db
-    environment:
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: postgres
-      POSTGRES_DB: claude_proxy
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-      - ../scripts/init-database.sql:/docker-entrypoint-initdb.d/01-init.sql
-    ports:
-      - "5432:5432"
-    healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U postgres"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
-
-  # Proxy Service - Built locally
-  proxy:
-    build:
-      context: ../
-      dockerfile: docker/proxy/Dockerfile
-    image: claude-nexus-proxy:local
-    container_name: claude-nexus-proxy
-    environment:
-      # Server configuration
-      PORT: 3000
-      HOST: 0.0.0.0
-      
-      # Claude API configuration
-      CREDENTIALS_DIR: /app/credentials
-      
-      # Database configuration
-      DATABASE_URL: postgresql://postgres:postgres@postgres:5432/claude_proxy
-      STORAGE_ENABLED: "true"
-      
-      # Telemetry
-      TELEMETRY_ENDPOINT: ${TELEMETRY_ENDPOINT}
-      
-      # Slack notifications (optional)
-      SLACK_WEBHOOK_URL: ${SLACK_WEBHOOK_URL}
-      SLACK_CHANNEL: ${SLACK_CHANNEL}
-      SLACK_USERNAME: ${SLACK_USERNAME:-Claude Nexus Proxy}
-      SLACK_ICON_EMOJI: ${SLACK_ICON_EMOJI:-:robot_face:}
-      SLACK_ENABLED: ${SLACK_ENABLED:-true}
-      
-      # Debug
-      DEBUG: ${DEBUG:-false}
-      
-      # Client authentication
-      ENABLE_CLIENT_AUTH: ${ENABLE_CLIENT_AUTH:-false}
-    volumes:
-      - ../credentials:/app/credentials
-      - ../client-setup:/app/client-setup:ro
-    ports:
-      - "3000:3000"
-    depends_on:
-      postgres:
-        condition: service_healthy
-    restart: unless-stopped
-    logging:
-      driver: "json-file"
-      options:
-        max-size: "10m"
-        max-file: "3"
-
-  # Dashboard Service - Built locally
-  dashboard:
-    build:
-      context: ../
-      dockerfile: docker/dashboard/Dockerfile
-    image: claude-nexus-dashboard:local
-    environment:
-      # Server configuration
-      PORT: 3001
-      HOST: 0.0.0.0
-      
-      # Dashboard authentication
-      DASHBOARD_API_KEY: key
-      
-      # Proxy API URL for data access
-      PROXY_API_URL: http://proxy:3000
-      
-      # Database configuration
-      DATABASE_URL: postgresql://postgres:postgres@postgres:5432/claude_proxy
-    ports:
-      - "3001:3001"
-    depends_on:
-      proxy:
-        condition: service_started
-    restart: unless-stopped
-
-  # Optional: pgAdmin for database management
-  # To enable: docker-compose up
-  pgadmin:
-    image: dpage/pgadmin4:latest
-    environment:
-      PGADMIN_DEFAULT_EMAIL: admin@example.com
-      PGADMIN_DEFAULT_PASSWORD: admin
-      PGADMIN_CONFIG_SERVER_MODE: 'False'
-      PGADMIN_CONFIG_MASTER_PASSWORD_REQUIRED: 'False'
-    ports:
-      - "5050:80"
-    depends_on:
-      - postgres
-
-  # Claude CLI Service - Run Claude CLI connected to the proxy
-  claude-cli:
-    build:
-      context: ./claude-cli
-      dockerfile: Dockerfile
-    container_name: claude-cli
-    environment:
-      # Proxy configuration for Claude Code
-      ANTHROPIC_BASE_URL: http://proxy:3000
-    volumes:
-      # Mount the entire project for Claude Code functionality
-      - ../:/workspace
-    depends_on:
-      proxy:
-        condition: service_started
-    stdin_open: true
-    tty: true
-    logging:
-      driver: "json-file"
-      options:
-        max-size: "10m"
-        max-file: "3"
-
-volumes:
-  postgres_data:
-  claude_data:
-
-networks:
-  default:
-    name: claude-nexus-network
-    external: true
