--- docker/docker-compose.override.yml
@@ -1,11 +1,0 @@
-# Docker Compose Override File
-# This file overrides settings for local development
-
-services:
-  proxy:
-    environment:
-      ENABLE_CLIENT_AUTH: "false"
-  
-  claude-cli:
-    environment:
-      ANTHROPIC_BASE_URL: "http://proxy:3000"
