+++ services/dashboard/src/layout/index.ts
@@ -1,0 +1,42 @@
+import { html } from 'hono/html'
+import { dashboardStyles } from './styles.js'
+
+/**
+ * Dashboard HTML layout template
+ */
+export const layout = (title: string, content: any, additionalScripts: string = '') => html`
+  <!DOCTYPE html>
+  <html lang="en">
+    <head>
+      <meta charset="UTF-8" />
+      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+      <title>${title} - Claude Nexus Dashboard</title>
+      <style>
+        ${dashboardStyles}
+      </style>
+      <link
+        rel="stylesheet"
+        href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.11.1/styles/github-dark.min.css"
+      />
+      <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.11.1/highlight.min.js"></script>
+      <script src="https://cdn.jsdelivr.net/npm/renderjson@1.4.0/renderjson.min.js"></script>
+      <script src="https://unpkg.com/htmx.org@1.9.10"></script>
+      ${additionalScripts}
+    </head>
+    <body>
+      <nav>
+        <div class="container">
+          <h1>Claude Nexus Dashboard</h1>
+          <div class="space-x-4">
+            <a href="/dashboard" class="text-sm text-blue-600">Dashboard</a>
+            <a href="/dashboard/requests" class="text-sm text-blue-600">Requests</a>
+            <a href="/dashboard/token-usage" class="text-sm text-blue-600">Token Usage</a>
+            <span class="text-sm text-gray-600" id="current-domain">All Domains</span>
+            <a href="/dashboard/logout" class="text-sm text-blue-600">Logout</a>
+          </div>
+        </div>
+      </nav>
+      <main class="container" style="padding: 2rem 1rem;">${content}</main>
+    </body>
+  </html>
+`
